@@ -3,7 +3,7 @@
   "version": "1.0.0",
   "description": "Identity and Access Management Dashboard for Auth0",
   "engines": {
-    "node": "~4.2.3"
+    "node": "~5.8.0"
   },
   "main": "server.js",
   "scripts": {
@@ -20,41 +22,55 @@
   "author": "Sandrino Di Mattia",
   "license": "MIT",
   "dependencies": {
-    "autoprefixer": "^6.2.3",
     "aws-sdk": "^2.2.26",
-    "axios": "^0.8.1",
+    "axios": "^0.9.1",
     "babel": "^6.3.26",
     "babel-core": "^6.3.26",
-    "babel-loader": "^6.2.0",
     "babel-plugin-transform-runtime": "^6.3.13",
     "babel-polyfill": "^6.3.14",
     "babel-preset-es2015": "^6.3.13",
-    "babel-preset-es2015-loose": "^6.1.4",
-    "babel-preset-react": "^6.3.13",
-    "babel-preset-react-hmre": "^1.0.1",
+    "babel-preset-es2015-loose": "^7.0.0",
     "babel-preset-stage-0": "^6.3.13",
     "babel-register": "^6.3.13",
     "babel-runtime": "^6.3.19",
     "bluebird": "^3.1.1",
     "body-parser": "^1.14.2",
-    "classnames": "^2.2.2",
-    "css-loader": "^0.23.1",
     "ejs": "^2.3.4",
-    "exports-loader": "^0.6.2",
     "express": "^4.13.3",
-    "extract-text-webpack-plugin": "^1.0.0",
-    "file-loader": "^0.8.5",
-    "history": "^1.17.0",
-    "immutable": "^3.7.6",
-    "imports-loader": "^0.6.5",
-    "jwt-decode": "^1.4.0",
-    "lodash": "^3.10.1",
-    "lowdb": "^0.11.2",
+    "lodash": "^4.6.1",
+    "lowdb": "^0.12.5",
     "moment": "^2.11.0",
     "mongo-getdb": "^2.0.0",
     "mongodb-uri": "^0.9.7",
     "morgan": "^1.6.1",
     "nconf": "^0.8.2",
+    "node-uuid": "^1.4.7",
+    "request": "^2.67.0",
+    "validate.js": "^0.9.0",
+    "webtask-tools": "^1.2.0",
+    "winston": "^2.1.1"
+  },
+  "devDependencies": {
+    "autoprefixer": "^6.2.3",
+    "babel-eslint": "^5.0.0",
+    "babel-loader": "^6.2.0",
+    "babel-preset-react": "^6.3.13",
+    "babel-preset-react-hmre": "^1.0.1",
+    "chai": "^3.5.0",
+    "classnames": "^2.2.2",
+    "css-loader": "^0.23.1",
+    "eslint": "^2.4.0",
+    "eslint-config-airbnb": "^6.1.0",
+    "eslint-plugin-babel": "^3.0.0",
+    "eslint-plugin-react": "^4.2.1",
+    "exports-loader": "^0.6.2",
+    "extract-text-webpack-plugin": "^1.0.0",
+    "file-loader": "^0.8.5",
+    "history": "^2.0.1",
+    "immutable": "^3.7.6",
+    "imports-loader": "^0.6.5",
+    "jwt-decode": "^1.4.0",
+    "mocha": "^2.4.5",
     "nodemon": "^1.8.1",
     "postcss-focus": "^1.0.0",
     "postcss-loader": "^0.8.0",
