@@ -1,8 +1,9 @@
 import ReactDOM from 'react-dom';
 import { Provider } from 'react-redux';
 import { Router, Route, IndexRedirect } from 'react-router';
-import { syncReduxAndRouter } from 'redux-simple-router';
-import createHistory from 'history/lib/createBrowserHistory';
+
+import { browserHistory } from 'react-router';
+import { routerMiddleware, syncHistoryWithStore } from 'react-router-redux';
 
 import { loadCredentials } from './actions/auth';
 
@@ -15,13 +16,14 @@ import UserContainer from './containers/User/UserContainer';
 import UsersContainer from './containers/Users/UsersContainer';
 import RolesContainer from './containers/Roles/RolesContainer';
 import GroupsContainer from './containers/Groups/GroupsContainer';
+import GroupContainer from './containers/Groups/GroupContainer';
 import PermissionsContainer from './containers/Permissions/PermissionsContainer';
 import ApplicationsContainer from './containers/Applications/ApplicationsContainer';
 
 import configureStore from './store/configureStore';
-const history = createHistory();
-const store = configureStore({ });
-syncReduxAndRouter(history, store);
+
+const store = configureStore([routerMiddleware(browserHistory)], { });
+const history = syncHistoryWithStore(browserHistory, store);
 
 // Fire first events.
 store.dispatch(loadCredentials());
@@ -34,7 +36,9 @@ ReactDOM.render(
         <IndexRedirect to="/users" />
         <Route path="applications" component={ApplicationsContainer} />
         <Route path="roles" component={RolesContainer} />
-        <Route path="groups" component={GroupsContainer} />
+        <Route path="groups" component={GroupsContainer}>
+          <Route path=":id" component={GroupContainer} />
+        </Route>
         <Route path="permissions" component={PermissionsContainer} />
         <Route path="logs" component={LogsContainer} />
         <Route path="users" component={UsersContainer}>
@@ -46,3 +50,6 @@ ReactDOM.render(
   </Provider>,
   document.getElementById('app')
 );
+
+  const showDevTools = require('./showDevTools');
+  showDevTools(store);
