@@ -58,7 +58,7 @@ export const SAVE_GROUP_REJECTED = 'SAVE_GROUP_REJECTED';
 export const SAVE_GROUP_FULFILLED = 'SAVE_GROUP_FULFILLED';
 
 export const CREATE_GROUP = 'CREATE_GROUP';
-export const EDIT_GROUP = 'SAVE_GROUP';
+export const EDIT_GROUP = 'EDIT_GROUP';
 export const CLEAR_GROUP = 'CLEAR_GROUP';
 
 export const REQUESTING_DELETE_GROUP = 'REQUESTING_DELETE_GROUP';
