@@ -4,7 +4,7 @@ import DockMonitor from 'redux-devtools-dock-monitor';
 
 export default createDevTools(
   <DockMonitor toggleVisibilityKey="ctrl-h"
-    changePositionKey="ctrl-q"
+    changePositionKey="ctrl-q" defaultSize={1} fluid={true}
   >
     <LogMonitor theme="tomorrow" />
   </DockMonitor>
